pvcviewer-controller: Add kustomization
Add kustomization for pvcviewer-controller.
To get this end result, we start with the original kustomizations generated by
kubebuilder and stored under components/pvcviewer-controller/config in
the kubeflow repo.
Afterwards we trim all unnecessary parts (webhooks, cert-manager) and
put everything in one kustomization.

We do this because kfctl currently doesn't work with multiple bases.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>