Use JSON patch to change dex Service
It seems that nodePort: null does not work as expected if a service already
exists and is a NodePort. It tries to apply the patch, and we end up with a
ClusterIP service with the *previous* nodePort value (also see [1]):

  The Service "dex" is invalid: spec.ports[0].nodePort: Forbidden: may not be used when `type` is 'ClusterIP'

The problem is that the null does not reach PATCH. It is interpreted
by kustomize SMP and it simply removes the nodePort from the Service definition
in base kustomization.

Changing SMP patch to JSON patch would work in a pure kustomize world. But
kfctl seems to interpret this as well and the *generated* dex-Service.yaml
does not include it.

[1] https://github.com/kubernetes/kubectl/issues/221

Signed-off-by: Dimitris Aragiorgis <dimara@arrikto.com>